requirements.txt
accelerate==1.10.1
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
diffusers==0.35.1
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
peft==0.17.1
pillow==11.0.0
protobuf==6.32.1
psutil==7.1.0
python-dotenv==1.1.1
python-telegram-bot==22.4
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==70.2.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.5.1+cu121
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing_extensions==4.12.2
urllib3==2.5.0
xformers==0.0.32.post2
zipp==3.23.0
